refactor(money): extract shared LegalLink to satisfy Sonar duplication gate
